Converted game.js to typescript and removed its responsibility to load and store the gamefile
